refactor: Update components layout
