+++ app/page.tsx
@@ -3,9 +3,9 @@ import Hero from "@/components/Hero";
 import WorkProjects from "@/components/WorkProjects";
 import Experience from "@/components/Experience";
 import Projects from "@/components/Projects";
-import { navItems } from "@/data";
 import BentoGrid from "@/components/BentoGrid";
-import Footer from "@/components/Footer";
+import Footer from "@/components/ui/Footer";
+import { navItems } from "@/data";
 
 export default function Home() {
   return (
