+++ components/ui/MagicButton.tsx
